+++ tests/integ/modin/series/test_dt_accessor.py
@@ -326,6 +326,36 @@ def test_dt_properties(property_name, freq):
 
 
 @pytest.mark.parametrize(
+    "property_name", ["days", "seconds", "microseconds", "nanoseconds"]
+)
+@sql_count_checker(query_count=1)
+def test_dt_timedelta_properties(property_name):
+    native_ser = native_pd.Series(
+        native_pd.TimedeltaIndex(
+            [
+                "1d",
+                "1h",
+                "60s",
+                "1s",
+                "800ms",
+                "5us",
+                "6ns",
+                "1d 3s",
+                "9m 15s 8us",
+                None,
+            ]
+        ),
+        index=[2, 6, 7, 8, 11, 16, 17, 20, 25, 27],
+        name="test",
+    )
+    snow_ser = pd.Series(native_ser)
+
+    eval_snowpark_pandas_result(
+        snow_ser, native_ser, lambda ser: getattr(ser.dt, property_name)
+    )
+
+
+@pytest.mark.parametrize(
     "data, data_type",
     [
         ([1, 2, 3, 4, 5], "int"),